Added band import data integrity check and associated tests. Made app use float64 for greater precision

+++ server/model/tax_calculator_test.go
@@ -26,7 +26,7 @@ func TestGetTaxDueForPropertyOfValue(t *testing.T) {
 	t.Run("calculateTaxDue should return zero if value of property < highest point on lowest band", func(t *testing.T) {
 		val, _ := tc.calculateTaxDue(0)
 
-		var want float32 = 0
+		var want float64 = 0
 		if val != want {
 			t.Error("Should have returned zero as property value within first band")
 		}
@@ -34,14 +34,14 @@ func TestGetTaxDueForPropertyOfValue(t *testing.T) {
 
 	t.Run("calculateTaxDue should correct amount if price within second band", func(t *testing.T) {
 		val, _ := tc.calculateTaxDue(200_000)
-		var want float32 = 1100.00
+		var want float64 = 1100.00
 		if val != want {
 			t.Error("Should have returned zero as property value within second band")
 		}
 	})
 	t.Run("calculateTaxDue should correct amount if price top of second band", func(t *testing.T) {
 		val, _ := tc.calculateTaxDue(250_000)
-		var want float32 = 2100.00
+		var want float64 = 2100.00
 		if val != want {
 			t.Error("Should have returned zero as property value top of second band")
 		}
@@ -49,7 +49,7 @@ func TestGetTaxDueForPropertyOfValue(t *testing.T) {
 
 	t.Run("calculateTaxDue should correct amount if price in top band", func(t *testing.T) {
 		val, _ := tc.calculateTaxDue(800_000)
-		var want float32 = 54_350.00
+		var want float64 = 54_350.00
 		if val != want {
 			t.Error("Should have returned zero as property value in top band")
 		}
